fix : kwh Double 변환 작업 완료

+++ src/main/java/live/ioteatime/frontservice/dto/DailyElectricityDto.java
@@ -13,6 +13,6 @@ import java.time.LocalDateTime;
 public class DailyElectricityDto {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime time;
-    private Long kwh;
+    private Double kwh;
     private Long bill;
 }
